Move the database from TiDB Cloud to a local SQLite file
TiDB Serverless' free tier had become the binding constraint on a dataset
that is a few thousand rows and about 5 MB. The whole thing now lives in one
SQLite file on the same box as the API, through Node's built-in `node:sqlite`
— no quota, no 63 ms round trip, and no native module to rebuild on deploy.

The public API shape does not change. Neither does `db/pool.ts`'s exported
surface (`q`, `qOne`, `tx`, `Q`) or the `[rows, fields]` tuple every call site
destructures, so the modules above it were left alone apart from the
statements that were genuinely MySQL-only:

  * 16 `ON DUPLICATE KEY UPDATE` upserts became `ON CONFLICT (...) DO UPDATE`
    with the conflict target spelled out;
  * the freight serial allocator's `LAST_INSERT_ID(x + 1)` trick became an
    `UPDATE ... RETURNING`, and the sync push now calls the one allocator
    rather than carrying a second copy of it;
  * `UPDATE ... JOIN (subquery)` in the cell-cleanup path became an
    IN-subquery;
  * `GREATEST` -> `MAX`, `INSERT IGNORE` -> `INSERT OR IGNORE`,
    `CAST(? AS JSON)` -> a plain parameter (JSON has no SQLite affinity, so
    the cast would have silently stored 0), `INTERVAL n DAY` -> cutoffs
    computed in JS and bound;
  * `UTC_TIMESTAMP(3)`/`NOW(3)` are kept as app-defined SQL functions rather
    than rewritten at ~50 call sites, returning exactly the
    'yyyy-MM-dd HH:mm:ss.SSS' the sync cursors are ordered by. `REGEXP` is
    registered too — SQLite parses the operator but ships no implementation.

Three things in pool.ts are load-bearing and commented as such: everything is
serialized through one FIFO lock (node:sqlite is synchronous and single
connection, so overlapping transactions would otherwise interleave into one
another), parameters are normalized before binding (mysql2 accepted booleans,
`undefined` and `Date`; node:sqlite throws on the first two and binds a Date
as NULL), and `IN (?)` array expansion is reimplemented because nine reporting
queries rely on it.

`FOR UPDATE` is dropped throughout. The lock is strictly stronger: one writer
at a time process-wide, which is what a single-dealer install already was.

MySQL's `ON UPDATE CURRENT_TIMESTAMP(3)` becomes a per-table trigger guarded
on `NEW.updated_at IS OLD.updated_at`, so a writer that sets the column itself
— the sync push and the backup import both do — still wins. Without that
guard the client's stored updated_at would never match the server's and every
pull would hand the row straight back as a change.

Also here, because moving off a managed database means owning what it did:

  * `jobs/backup.ts` takes a `VACUUM INTO` snapshot at boot and daily, keeping
    SQLITE_BACKUP_KEEP of them. Deliberately not behind ENABLE_JOBS — that
    flag gates the job that deletes rows, and tying backups to it is how you
    end up with none.
  * the integration suite is no longer gated on TEST_DATABASE_URL reaching a
    live TiDB, which meant in practice it never ran. It makes its own throwaway
    file per run, so all 18 tests are part of `npm test` (90 passing). Doing
    that surfaced the `UPDATE ... JOIN` above, which nothing else would have.
    The suite also predated mandatory email verification and refresh-rotation
    grace, so its expectations are brought up to the current behaviour.

The data was copied by scripts/migrate_tidb_to_sqlite.ts, kept in the repo as
the record of how: 8,395 rows, every table's count verified against the source
and `PRAGMA foreign_key_check` clean. The TiDB database is untouched.

+++ src/modules/masters/routes.ts
@@ -51,7 +51,7 @@ export const MASTER_DEFS: MasterDef[] = [
     }),
     // §3.2: order ASC, numeric code (non-numeric sorts last at rank 1<<20), name CI.
     orderBy: `ORDER BY sort_order ASC,
-      CASE WHEN code IS NULL OR code NOT REGEXP '^[0-9]+$' THEN 1048576 ELSE CAST(code AS DECIMAL(20,0)) END ASC,
+      CASE WHEN code IS NULL OR code NOT REGEXP '^[0-9]+$' THEN 1048576 ELSE CAST(code AS INTEGER) END ASC,
       LOWER(name) ASC`,
   },
   {
